fixed scoring/completeness on hazards
added ability for hazards to declare their complete-ness to update score menu. changed level editior, needs to be re-updated with resources. fixed pause menu.

+++ VR-CTS/Assets/Scripts/ObjectivesListUI.cs
@@ -9,14 +9,14 @@ public class ObjectivesListUI : MonoBehaviour
 	private Text[] hazardCountTexts;
 	public Transform anchor;
 	public GameObject hazardText;
-	public int textOffset;
+	public float textOffset;
 
 	private int[] hazardCompleted;
 	public int[] hazardTotals;
 
 	public void RecordHazards(int[] hazardCountArray) {
         GameObject curObject;
-        int curOffset = 0;
+        float curOffset = 0;
         int count = hazardCountArray.Length;
         hazardCountTexts = new Text[count];
         hazardCompleted = new int[count];
